+++ packages/frontend-embed/src/theme.ts
@@ -6,8 +6,8 @@
 // TODO: (可能な部分を)sharedに抽出して frontend と共通化
 
 import tinycolor from 'tinycolor2';
-import lightTheme from '@@/themes/_light.json5';
-import darkTheme from '@@/themes/_dark.json5';
+import lightTheme from '@@/themes/_light.aison';
+import darkTheme from '@@/themes/_dark.aison';
 import type { BundledTheme } from 'shiki/themes';
 
 export type Theme = {