+++ packages/frontend/src/components/MkExtensionInstaller.stories.impl.ts
@@ -5,7 +5,7 @@
 
 import type { StoryObj } from '@storybook/vue3';
 import MkExtensionInstaller from './MkExtensionInstaller.vue';
-import lightTheme from '@@/themes/_light.json5';
+import lightTheme from '@@/themes/_light.aison';
 
 export const Plugin = {
 	render(args) {