+++ packages/frontend/src/pages/settings/theme.install.vue
@@ -5,7 +5,7 @@ SPDX-License-Identifier: AGPL-3.0-only
 
 <template>
 <div class="_gaps_m">
-	<MkCodeEditor v-model="installThemeCode" lang="json5">
+	<MkCodeEditor v-model="installThemeCode" lang="js">
 		<template #label>{{ i18n.ts._theme.code }}</template>
 	</MkCodeEditor>